Updated all python scripts in scripts folder
- Made updates to create_stack_dash, create_stack, create_dashboard, and delete_stack.
- create_stack_dash now uses the config.env file to get arguments/parameters, it no longer relies on a text file (ex: args.txt).
- Feature to delete individual stacks created per run has been implemented, deletion via prefix is available as an option.
- Instructions updated to explain new usage

+++ jmeter-icap-poc/scripts/delete_stack.py
@@ -1,41 +1,45 @@
 import boto3
 from datetime import timedelta, datetime, timezone
-import os
 import argparse
+from create_stack_dash import Config
 
 
-def get_configuration(key):
-    # Load configuration
-    try:
+def main(config):
+    profile_name = config.aws_profile_name
 
-        if os.path.exists("config.env"):
-            with open("config.env") as f:
-                config = f.readlines()
-            configuration = dict(c.strip().split("=") for c in config)
-            return configuration.get(key)
-        else:
-            return os.getenv(key.upper())
-    except Exception as e:
-        print("Please create config.env file similar to config.env.sample or set environment variables for all variables in config.env.sample file")
-        print(str(e))
-        raise
-
-
-def main():
-    profile_name = get_configuration("aws_profile_name")
-    parser = argparse.ArgumentParser(description='Create cloudformation stack to deploy ASG.')
-    parser.add_argument('--stack_name', '-n', default="",
-                        help='Cloudformation stack name of stack to be deleted')
-
-    args = parser.parse_args()
-    stack_name = args.stack_name
+    prefix = config.prefix + "-" if config.prefix not in ["", None] else config.prefix
+    min_age = config.min_age
+    stack_name = config.stack_name
 
     session = boto3.session.Session(profile_name=profile_name)
     client = session.client("cloudformation")
-    
-    print("deleting stack named: %s" % (stack_name))
-    client.delete_stack(StackName=stack_name)
+
+    now = datetime.now(timezone.utc)
+    past_time = now - timedelta(minutes=min_age)
+    stacks_list = client.list_stacks(StackStatusFilter=["CREATE_COMPLETE", "CREATE_FAILED"])
+
+    if config.prefix_based_delete:
+        print("finding the stack names with prefix %s" % prefix)
+        for s in stacks_list["StackSummaries"]:
+            stack_name = s["StackName"]
+            if not stack_name.startswith(prefix + "aws-jmeter-test-engine-"):
+                continue
+
+            # This will delete stacks that are older than given age.
+            creation_time = s["CreationTime"]
+            if creation_time <= past_time:
+                print("deleting stack %s" % (stack_name))
+                client.delete_stack(StackName=stack_name)
+
+        print("finding stack with name %s" % stack_name)
+        for s in stacks_list["StackSummaries"]:
+            stack_name = s["StackName"]
+            if not stack_name.startswith(prefix + "aws-jmeter-test-engine-"):
+                continue
+    else:
+        print("deleting stack named: %s" % (stack_name))
+        client.delete_stack(StackName=stack_name)
 
 
 if __name__ == "__main__":
-    main()
+    main(Config)
